moved methods to win_32 adapter and added spec files
